Changing a cast from unsigned to uint64_t, should be NFC in practice.
llvm-svn: 224249

+++ llvm/lib/MC/MCAsmStreamer.cpp
@@ -683,7 +683,7 @@ void MCAsmStreamer::EmitValueImpl(const MCExpr *Value, unsigned Size,
       // emission domain.  This produces nicer output and silences potential
       // truncation warnings when round tripping through another assembler.
       uint64_t Shift = 64 - EmissionSize * 8;
-      assert(Shift < static_cast<unsigned>(
+      assert(Shift < static_cast<uint64_t>(
                          std::numeric_limits<unsigned long long>::digits) &&
              "undefined behavior");
       ValueToEmit &= ~0ULL >> Shift;
